Karl branch

+++ DressMe/src/br/com/vsetsistemas/model/Cargo.java
@@ -2,21 +2,21 @@ package br.com.vsetsistemas.model;
 
 public class Cargo {
 
-	private int id;
+	private long id;
 	private String descricao;
 	private boolean status;
 	
-	public Cargo(int id, String descricao, boolean status) {
+	public Cargo(long id, String descricao, boolean status) {
 		this.id = id;
 		this.descricao = descricao;
 		this.status = status;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
