+++ DressMe/src/br/com/vsetsistemas/model/Categoria.java
@@ -2,10 +2,10 @@ package br.com.vsetsistemas.model;
 
 public class Categoria {
 
-	private int id;
+	private long id;
 	private String descricao;
 	
-	public Categoria(int id, String descricao) {
+	public Categoria(long id, String descricao) {
 		this.id = id;
 		this.descricao = descricao;
 	}
@@ -13,11 +13,11 @@ public class Categoria {
 	public Categoria() {
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
